Renamed target finder class and added to optional CMake component
